requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
redis
rq
boto3
